+++ Editor/Effectors/IRigEffectorOverlay.cs
@@ -1,7 +1,10 @@
+using System;
+
 namespace UnityEditor.Animations.Rigging
 {
-    interface IRigEffectorOverlay
+    interface IRigEffectorOverlay : IDisposable
     {
+        bool IsValid();
         void OnSceneGUIOverlay();
     }
 }