+++ Runtime/AnimationRig/Constraints/BlendConstraint.cs
@@ -74,7 +74,7 @@ namespace UnityEngine.Animations.Rigging
     /// Blend constraint.
     /// </summary>
     [DisallowMultipleComponent, AddComponentMenu("Animation Rigging/Blend Constraint")]
-    [HelpURL("https://docs.unity3d.com/Packages/com.unity.animation.rigging@1.2/manual/constraints/BlendConstraint.html")]
+    [HelpURL("https://docs.unity3d.com/Packages/com.unity.animation.rigging@1.3/manual/constraints/BlendConstraint.html")]
     public class BlendConstraint : RigConstraint<
         BlendConstraintJob,
         BlendConstraintData,